Moved impl's to seperate files for every type
