+++ crates/duckdb/src/types/tests.rs
@@ -1,4 +1,5 @@
 use crate::environment::{Environment, StorageLocation};
+use crate::logical_type::LogicalTypeID;
 
 use super::*;
